Add simple store test for JS decorated class
